+++ src/directives/index.ts
@@ -1,4 +1,5 @@
 export * from './normalStateTemplate/normalStateTemplate.directive';
+export * from './normalStateTagTemplate/normalStateTagTemplate.directive';
 export * from './optionGroupTemplate/optionGroupTemplate.directive';
 export * from './optionTemplate/optionTemplate.directive';
 export * from './selectAbsolute/selectAbsolute.directive';