+++ src/interfaces/normalStateContext/normalStateContext.interface.ts
@@ -3,10 +3,10 @@ import {NormalState} from '../plugins';
 /**
  * Context for template that is used within normal state plugin for rendering content
  */
-export interface NormalStateContext
+export interface NormalStateContext<TValue = unknown>
 {
     /**
      * Instance of plugin itself
      */
-    $implicit: NormalState;
+    $implicit: NormalState<TValue>;
 }