+++ src/interfaces/popupContext/popupContext.interface.ts
@@ -1,18 +1,18 @@
 import {Popup} from '../plugins';
-import {SelectOption} from '../selectOption/selectOption.interface';
+import {SelectOptionState} from '../selectOptionState/selectOptionState.interface';
 
 /**
  * Context for template that is used within popup plugin for rendering option
  */
-export interface PopupContext
+export interface PopupContext<TValue = unknown>
 {
     /**
      * Instance of plugin itself
      */
-    $implicit: SelectOption;
+    $implicit: SelectOptionState<TValue>;
 
     /**
      * Instance of plugin itself
      */
-    popup: Popup;
+    popup: Popup<TValue>;
 }